Consolidate expression string tests and private helpers
