--- internal/exprtest/roundtrip_test.go
@@ -1,32 +1,0 @@
-// Copyright 2026 Dolthub, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package exprtest
-
-import (
-	"testing"
-
-	"github.com/dolthub/vitess/go/vt/sqlparser"
-	"github.com/stretchr/testify/require"
-
-	"github.com/dolthub/go-mysql-server/sql/expression"
-	"github.com/dolthub/go-mysql-server/sql/types"
-)
-
-func TestParseExpressionReturnsTypedValue(t *testing.T) {
-	original := expression.NewLiteral("value", types.Text)
-	parsed := RequireExpression(t, ParseExpression(t, original))
-	value := requireSQLValue(t, parsed, sqlparser.StrVal)
-	require.Equal(t, original.Value(), string(value.Val))
-}
