--- sql/expression/alias_test.go
@@ -1,37 +1,0 @@
-// Copyright 2026 Dolthub, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package expression
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/require"
-
-	"github.com/dolthub/go-mysql-server/internal/exprtest"
-	"github.com/dolthub/go-mysql-server/sql"
-	"github.com/dolthub/go-mysql-server/sql/types"
-)
-
-func TestAliasReferenceString(t *testing.T) {
-	expr := NewAliasReference("alias_name")
-	require.Equal(t, "alias_name", expr.String())
-	exprtest.AssertColumnRoundTrip(t, expr)
-}
-
-func TestAliasString(t *testing.T) {
-	expr := NewAlias(sql.NewEmptyContext(), "alias name", NewLiteral(42, types.Int64))
-	require.Equal(t, "42 as `alias name`", expr.String())
-	exprtest.AssertAliasRoundTrip(t, expr)
-}
