+++ sql/expression/function/aggregation/avg_test.go
@@ -19,25 +19,11 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	"github.com/dolthub/go-mysql-server/internal/exprtest"
 	"github.com/dolthub/go-mysql-server/sql"
 	"github.com/dolthub/go-mysql-server/sql/expression"
 	"github.com/dolthub/go-mysql-server/sql/types"
 	_ "github.com/dolthub/go-mysql-server/sql/variables"
 )
-
-func TestAvg_String(t *testing.T) {
-	require := require.New(t)
-
-	avg := NewAvg(expression.NewGetField(0, types.Int32, "col1", true))
-	require.Equal("AVG(col1)", avg.String())
-
-	windowed := avg.WithWindow(sql.NewEmptyContext(), &sql.WindowDefinition{
-		PartitionBy: []sql.Expression{expression.NewGetField(1, types.Int32, "col2", true)},
-	})
-	require.Equal("AVG(col1) over ( partition by col2)", windowed.String())
-	exprtest.AssertFunctionRoundTrip(t, windowed.(sql.FunctionExpression))
-}
 
 func TestAvg_Float64(t *testing.T) {
 	require := require.New(t)
