+++ sql/expression/function/aggregation/bitwise_test.go
@@ -19,21 +19,10 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	"github.com/dolthub/go-mysql-server/internal/exprtest"
 	"github.com/dolthub/go-mysql-server/sql"
 	"github.com/dolthub/go-mysql-server/sql/expression"
 	"github.com/dolthub/go-mysql-server/sql/types"
 )
-
-func TestBitAnd_String(t *testing.T) {
-	assert := require.New(t)
-	m := NewBitAnd(expression.NewGetField(0, types.Int32, "field", true))
-	assert.Equal("BIT_AND(field)", m.String())
-
-	windowed := m.WithWindow(sql.NewEmptyContext(), &sql.WindowDefinition{})
-	assert.Equal("BIT_AND(field) over ()", windowed.String())
-	exprtest.AssertFunctionRoundTripAs(t, windowed.(sql.FunctionExpression), "BIT_AND")
-}
 
 func TestBitAnd_Eval_Int(t *testing.T) {
 	assert := require.New(t)
@@ -111,16 +100,6 @@ func TestBitAnd_Eval_Empty(t *testing.T) {
 	assert.Equal(^uint64(0), v)
 }
 
-func TestBitOr_String(t *testing.T) {
-	assert := require.New(t)
-	m := NewBitOr(expression.NewGetField(0, types.Int32, "field", true))
-	assert.Equal("BIT_OR(field)", m.String())
-
-	windowed := m.WithWindow(sql.NewEmptyContext(), &sql.WindowDefinition{})
-	assert.Equal("BIT_OR(field) over ()", windowed.String())
-	exprtest.AssertFunctionRoundTripAs(t, windowed.(sql.FunctionExpression), "BIT_OR")
-}
-
 func TestBitOr_Eval_Int(t *testing.T) {
 	assert := require.New(t)
 	ctx := sql.NewEmptyContext()
@@ -195,16 +174,6 @@ func TestBitOr_Eval_Empty(t *testing.T) {
 	v, err := b.Eval(ctx)
 	assert.NoError(err)
 	assert.Equal(uint64(0), v)
-}
-
-func TestBitXor_String(t *testing.T) {
-	assert := require.New(t)
-	m := NewBitXor(expression.NewGetField(0, types.Int32, "field", true))
-	assert.Equal("BIT_XOR(field)", m.String())
-
-	windowed := m.WithWindow(sql.NewEmptyContext(), &sql.WindowDefinition{})
-	assert.Equal("BIT_XOR(field) over ()", windowed.String())
-	exprtest.AssertFunctionRoundTripAs(t, windowed.(sql.FunctionExpression), "BIT_XOR")
 }
 
 func TestBitXor_Eval_Int(t *testing.T) {
