+++ sql/expression/function/aggregation/last_test.go
@@ -19,7 +19,6 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	"github.com/dolthub/go-mysql-server/internal/exprtest"
 	"github.com/dolthub/go-mysql-server/sql"
 	"github.com/dolthub/go-mysql-server/sql/expression"
 	"github.com/dolthub/go-mysql-server/sql/types"
@@ -43,11 +42,4 @@ func TestLast(t *testing.T) {
 			require.Equal(t, tt.expected, result)
 		})
 	}
-}
-
-func TestLastString(t *testing.T) {
-	expr := NewLast(expression.NewGetField(0, types.Text, "value", false)).
-		WithWindow(sql.NewEmptyContext(), &sql.WindowDefinition{})
-	require.Equal(t, "LAST_VALUE(value) over ()", expr.String())
-	exprtest.AssertFunctionRoundTripAs(t, expr.(sql.FunctionExpression), "LAST_VALUE")
 }
