+++ sql/expression/function/aggregation/std_test.go
@@ -20,20 +20,12 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	"github.com/dolthub/go-mysql-server/internal/exprtest"
 	"github.com/dolthub/go-mysql-server/sql"
 	"github.com/dolthub/go-mysql-server/sql/expression"
 )
 
 func isFloatEqual(a, b float64) bool {
 	return math.Abs(a-b) < 1e-9
-}
-
-func TestStdDevPopString(t *testing.T) {
-	expr := NewStdDevPop(expression.NewGetField(0, nil, "value", false)).
-		WithWindow(sql.NewEmptyContext(), &sql.WindowDefinition{})
-	require.Equal(t, "STDDEV_POP(value) over ()", expr.String())
-	exprtest.AssertFunctionRoundTripAs(t, expr.(sql.FunctionExpression), "STDDEV_POP")
 }
 
 func TestStd(t *testing.T) {
@@ -188,20 +180,6 @@ func TestStdSamp(t *testing.T) {
 	}
 }
 
-func TestStdDevSampString(t *testing.T) {
-	expr := NewStdDevSamp(expression.NewGetField(0, nil, "value", false)).
-		WithWindow(sql.NewEmptyContext(), &sql.WindowDefinition{})
-	require.Equal(t, "STDDEV_SAMP(value) over ()", expr.String())
-	exprtest.AssertFunctionRoundTripAs(t, expr.(sql.FunctionExpression), "STDDEV_SAMP")
-}
-
-func TestVarPopString(t *testing.T) {
-	expr := NewVarPop(expression.NewGetField(0, nil, "value", false)).
-		WithWindow(sql.NewEmptyContext(), &sql.WindowDefinition{})
-	require.Equal(t, "VAR_POP(value) over ()", expr.String())
-	exprtest.AssertFunctionRoundTripAs(t, expr.(sql.FunctionExpression), "VAR_POP")
-}
-
 func TestVariance(t *testing.T) {
 	sum := NewVarPop(expression.NewGetField(0, nil, "", false))
 
@@ -276,13 +254,6 @@ func TestVariance(t *testing.T) {
 			require.True(isFloatEqual(tt.expected.(float64), result.(float64)))
 		})
 	}
-}
-
-func TestVarSampString(t *testing.T) {
-	expr := NewVarSamp(expression.NewGetField(0, nil, "value", false)).
-		WithWindow(sql.NewEmptyContext(), &sql.WindowDefinition{})
-	require.Equal(t, "VAR_SAMP(value) over ()", expr.String())
-	exprtest.AssertFunctionRoundTripAs(t, expr.(sql.FunctionExpression), "VAR_SAMP")
 }
 
 func TestVarSamp(t *testing.T) {
