+++ sql/expression/function/aggregation/sum_test.go
@@ -19,17 +19,9 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	"github.com/dolthub/go-mysql-server/internal/exprtest"
 	"github.com/dolthub/go-mysql-server/sql"
 	"github.com/dolthub/go-mysql-server/sql/expression"
 )
-
-func TestSumString(t *testing.T) {
-	expr := NewSum(expression.NewGetField(0, nil, "value", false)).
-		WithWindow(sql.NewEmptyContext(), &sql.WindowDefinition{})
-	require.Equal(t, "SUM(value) over ()", expr.String())
-	exprtest.AssertFunctionRoundTrip(t, expr.(sql.FunctionExpression))
-}
 
 func TestSum(t *testing.T) {
 	sum := NewSum(expression.NewGetField(0, nil, "", false))
