+++ sql/expression/function/functionregistry_test.go
@@ -19,12 +19,9 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	"github.com/dolthub/go-mysql-server/internal/exprtest"
 	"github.com/dolthub/go-mysql-server/sql"
 	"github.com/dolthub/go-mysql-server/sql/expression"
 	"github.com/dolthub/go-mysql-server/sql/expression/function"
-	"github.com/dolthub/go-mysql-server/sql/expression/function/spatial"
-	"github.com/dolthub/go-mysql-server/sql/types"
 )
 
 func TestFunctionRegistry(t *testing.T) {
@@ -62,33 +59,4 @@ func TestFunctionRegistryMissingFunction(t *testing.T) {
 	f, ok := reg.Function(sql.NewEmptyContext(), "", "func")
 	require.False(ok)
 	require.Nil(f)
-}
-
-func TestMultiLineStringFromTextRegistryString(t *testing.T) {
-	ctx := sql.NewEmptyContext()
-	reg := function.NewRegistry()
-	reg.Register(function.BuiltIns...)
-	fn, ok := reg.Function(ctx, "", "st_multilinestringfromtext")
-	require.True(t, ok)
-	expr, err := fn.NewInstance(ctx, []sql.Expression{
-		expression.NewLiteral("MULTILINESTRING((1 2, 3 4))", types.Text),
-	})
-	require.NoError(t, err)
-	require.IsType(t, &spatial.MLineFromText{}, expr)
-	require.Equal(t, "st_mlinefromtext('MULTILINESTRING((1 2, 3 4))')", expr.String())
-	exprtest.AssertFunctionRoundTrip(t, expr.(sql.FunctionExpression))
-}
-
-func TestSTEqualsRegistryString(t *testing.T) {
-	ctx := sql.NewEmptyContext()
-	reg := function.NewRegistry()
-	reg.Register(function.BuiltIns...)
-	fn, ok := reg.Function(ctx, "", "st_equals")
-	require.True(t, ok)
-	point := expression.NewLiteral(types.Point{X: 1, Y: 2}, types.PointType{})
-	expr, err := fn.NewInstance(ctx, []sql.Expression{point, point})
-	require.NoError(t, err)
-	require.IsType(t, &spatial.STEquals{}, expr)
-	require.Equal(t, "ST_EQUALS(ST_GeomFromWKB(0x0101000000000000000000F03F0000000000000040, 0), ST_GeomFromWKB(0x0101000000000000000000F03F0000000000000040, 0))", expr.String())
-	exprtest.AssertFunctionRoundTrip(t, expr.(sql.FunctionExpression))
 }
