+++ sql/expression/function/hash_test.go
@@ -20,7 +20,6 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	"github.com/dolthub/go-mysql-server/internal/exprtest"
 	"github.com/dolthub/go-mysql-server/sql"
 	"github.com/dolthub/go-mysql-server/sql/expression"
 	"github.com/dolthub/go-mysql-server/sql/types"
@@ -393,10 +392,6 @@ func TestUncompress(t *testing.T) {
 }
 
 func TestUncompressedLength(t *testing.T) {
-	stringExpr := NewUncompressedLength(sql.NewEmptyContext(), expression.NewLiteral("value", types.Text))
-	require.Equal(t, "uncompressed_length('value')", stringExpr.String())
-	exprtest.AssertFunctionRoundTrip(t, stringExpr.(sql.FunctionExpression))
-
 	tests := []struct {
 		val sql.Expression
 		exp interface{}
@@ -452,10 +447,6 @@ func TestUncompressedLength(t *testing.T) {
 }
 
 func TestValidatePasswordStrength(t *testing.T) {
-	stringExpr := NewValidatePasswordStrength(sql.NewEmptyContext(), expression.NewLiteral("value", types.Text))
-	require.Equal(t, "validate_password_strength('value')", stringExpr.String())
-	exprtest.AssertFunctionRoundTrip(t, stringExpr.(sql.FunctionExpression))
-
 	tests := []struct {
 		val sql.Expression
 		exp interface{}
