+++ sql/expression/function/json/jsontests/json_function_test.go
@@ -17,11 +17,9 @@ package jsontests
 import (
 	"testing"
 
-	"github.com/dolthub/vitess/go/vt/sqlparser"
 	"github.com/stretchr/testify/require"
 	"gopkg.in/src-d/go-errors.v1"
 
-	"github.com/dolthub/go-mysql-server/internal/exprtest"
 	"github.com/dolthub/go-mysql-server/sql"
 	"github.com/dolthub/go-mysql-server/sql/expression"
 	"github.com/dolthub/go-mysql-server/sql/expression/function/json"
@@ -112,36 +110,6 @@ func TestJsonValue(t *testing.T) {
 			RunJsonValueTests(t, format.prepareFunc)
 		})
 	}
-}
-
-func TestJsonValueString(t *testing.T) {
-	ctx := sql.NewEmptyContext()
-	doc := expression.NewGetField(0, types.JSON, "doc", false)
-	path := expression.NewLiteral("$.a", types.Text)
-
-	signed, err := json.NewJsonValue(ctx, doc, path, expression.NewLiteral(int64(0), types.Int64))
-	require.NoError(t, err)
-	require.Equal(t, "json_value(doc, '$.a', 'signed')", signed.String())
-	parsedSigned := exprtest.RequireFunction(t, exprtest.ParseExpression(t, signed))
-	signedValue := signed.(*json.JsonValue)
-	require.Equal(t, signedValue.FunctionName(), parsedSigned.Name.Lowered())
-	require.Len(t, parsedSigned.Exprs, 3)
-	exprtest.AssertExpressionValue(t, exprtest.RequireFunctionArgument(t, parsedSigned, 0), signedValue.JSON)
-	exprtest.AssertExpressionValue(t, exprtest.RequireFunctionArgument(t, parsedSigned, 1), signedValue.Path)
-	parsedType := exprtest.RequireFunctionArgument(t, parsedSigned, 2).(*sqlparser.SQLVal)
-	require.Equal(t, sqlparser.StrVal, parsedType.Type)
-	require.True(t, types.IsSigned(signedValue.Typ))
-	require.Equal(t, "signed", string(parsedType.Val))
-
-	defaultType, err := json.NewJsonValue(ctx, doc, path)
-	require.NoError(t, err)
-	require.Equal(t, "json_value(doc, '$.a')", defaultType.String())
-	parsedDefault := exprtest.RequireFunction(t, exprtest.ParseExpression(t, defaultType))
-	defaultValue := defaultType.(*json.JsonValue)
-	require.Equal(t, defaultValue.FunctionName(), parsedDefault.Name.Lowered())
-	require.Len(t, parsedDefault.Exprs, 2)
-	exprtest.AssertExpressionValue(t, exprtest.RequireFunctionArgument(t, parsedDefault, 0), defaultValue.JSON)
-	exprtest.AssertExpressionValue(t, exprtest.RequireFunctionArgument(t, parsedDefault, 1), defaultValue.Path)
 }
 
 func TestJsonContainsPath(t *testing.T) {
