+++ sql/expression/unresolved_test.go
@@ -19,7 +19,6 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	"github.com/dolthub/go-mysql-server/internal/exprtest"
 	"github.com/dolthub/go-mysql-server/sql"
 )
 
@@ -31,28 +30,4 @@ func TestUnresolvedExpression(t *testing.T) {
 	require.NotNil(o)
 	o = NewNot(e)
 	require.NotNil(o)
-}
-
-func TestUnresolvedColumnString(t *testing.T) {
-	tests := []struct {
-		expr     *UnresolvedColumn
-		expected string
-	}{
-		{NewUnresolvedColumn("normal_name"), "normal_name"},
-		{NewUnresolvedQualifiedColumn("table_name", "column_name"), "table_name.column_name"},
-	}
-
-	for _, test := range tests {
-		require.Equal(t, test.expected, test.expr.String())
-		exprtest.AssertColumnRoundTrip(t, test.expr)
-	}
-}
-
-func TestUnresolvedFunctionString(t *testing.T) {
-	expr := NewUnresolvedFunction("function name", false, nil)
-	require.Equal(t, "`function name`()", expr.String())
-	parsed := exprtest.RequireFunction(t, exprtest.ParseExpression(t, expr))
-	require.Equal(t, expr.Name(), parsed.Name.String())
-	require.Equal(t, len(expr.Arguments), len(parsed.Exprs))
-	require.Nil(t, parsed.Over)
 }
