--- sql/expression/wrapper_test.go
@@ -1,47 +1,0 @@
-// Copyright 2026 Dolthub, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package expression
-
-import (
-	"testing"
-
-	"github.com/dolthub/vitess/go/vt/sqlparser"
-	"github.com/stretchr/testify/require"
-
-	"github.com/dolthub/go-mysql-server/internal/exprtest"
-	"github.com/dolthub/go-mysql-server/sql/types"
-)
-
-func TestWrapperString(t *testing.T) {
-	tests := []struct {
-		expr     *Wrapper
-		expected string
-	}{
-		{WrapExpression(nil), "NULL"},
-		{WrapExpression(NewLiteral(42, types.Int64)), "(42)"},
-	}
-
-	for _, test := range tests {
-		require.Equal(t, test.expected, test.expr.String())
-		parsed := exprtest.RequireExpression(t, exprtest.ParseExpression(t, test.expr))
-		if test.expr.Unwrap() == nil {
-			require.IsType(t, &sqlparser.NullVal{}, parsed)
-		} else {
-			parenthesized, ok := parsed.(*sqlparser.ParenExpr)
-			require.True(t, ok)
-			exprtest.AssertExpressionValue(t, parenthesized.Expr, test.expr.Unwrap())
-		}
-	}
-}
