--- sql/plan/alter_table_test.go
@@ -1,34 +1,0 @@
-// Copyright 2026 Dolthub, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package plan
-
-import (
-	"testing"
-
-	"github.com/dolthub/vitess/go/vt/sqlparser"
-	"github.com/stretchr/testify/require"
-
-	"github.com/dolthub/go-mysql-server/internal/exprtest"
-	"github.com/dolthub/go-mysql-server/sql"
-	"github.com/dolthub/go-mysql-server/sql/types"
-)
-
-func TestColDefaultExpressionString(t *testing.T) {
-	expr := ColDefaultExpression{Column: &sql.Column{Name: "odd` name", Type: types.Int64}}
-	require.Equal(t, "DEFAULT(`odd`` name`)", expr.String())
-	parsed, ok := exprtest.RequireExpression(t, exprtest.ParseExpression(t, expr)).(*sqlparser.Default)
-	require.True(t, ok)
-	require.Equal(t, expr.Column.Name, parsed.ColName)
-}
